requirements.txt
aiofiles==24.1.0
beautifulsoup4==4.13.3
faiss-cpu==1.11.0
fastapi==0.115.12
jinja2==3.1.6
lxml==6.0.2
matplotlib==3.10.3
numpy==2.3.0
ollama==0.6.0
pandas==2.3.0
pydantic==2.11.5
requests==2.32.3
rich==14.0.0
scikit-learn==1.7.0
sentence-transformers==3.0.1
torch==2.7.1
transformers==4.52.4
uvicorn==0.34.3